added authorization and authentication

+++ codes/Routers/posts.py
@@ -9,6 +9,8 @@ import models
 from database import get_db
 from schemas import PostCreate, PostResponse, PostUpdate
 
+from auth import CurrentUser
+
 router = APIRouter()
 
 
@@ -29,22 +31,16 @@ async def get_posts(db: Annotated[AsyncSession, Depends(get_db)]):
     response_model=PostResponse,
     status_code=status.HTTP_201_CREATED,
 )
-async def create_post(post: PostCreate, db: Annotated[AsyncSession, Depends(get_db)]):
-    result = await db.execute(
-        Select(models.User).where(models.User.id == post.user_id),
-    )
-    user = result.scalars().first()
-
-    if not user:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail="User not found",
-        )
+async def create_post(
+    post: PostCreate,
+    current_user: CurrentUser,
+    db: Annotated[AsyncSession, Depends(get_db)],
+):
 
     new_post = models.Post(
         title=post.title,
         content=post.content,
-        user_id=post.user_id,
+        user_id=current_user.id,
     )
 
     db.add(new_post)
@@ -72,7 +68,10 @@ async def get_post(post_id: int, db: Annotated[AsyncSession, Depends(get_db)]):
 
 @router.put("/{post_id}", response_model=PostResponse)
 async def update_post_full(
-    post_id: int, db: Annotated[AsyncSession, Depends(get_db)], post_data: PostCreate
+    post_id: int,
+    db: Annotated[AsyncSession, Depends(get_db)],
+    current_user: CurrentUser,
+    post_data: PostCreate,
 ):
     result = await db.execute(Select(models.Post).where(models.Post.id == post_id))
     post = result.scalars().first()
@@ -81,20 +80,24 @@ async def update_post_full(
             status_code=status.HTTP_404_NOT_FOUND, detail="Post not found"
         )
 
-    if post.user_id != post_data.user_id:
-        result = await db.execute(
-            Select(models.User).where(models.User.id == post_data.user_id)
+    # if post.user_id != post_data.user_id:
+    #     result = await db.execute(
+    #         Select(models.User).where(models.User.id == post_data.user_id)
+    #     )
+    #     user = result.scalars().first()
+    #     if not user:
+    #         raise HTTPException(
+    #             status_code=status.HTTP_404_NOT_FOUND,
+    #             detail="User not found",
+    #         )
+    if post.user_id != current_user.id:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="Not authorised to update this post",
         )
-        user = result.scalars().first()
-        if not user:
-            raise HTTPException(
-                status_code=status.HTTP_404_NOT_FOUND,
-                detail="User not found",
-            )
 
     post.title = post_data.title
     post.content = post_data.content
-    post.user_id = post_data.user_id
 
     await db.commit()
     await db.refresh(post)
@@ -103,13 +106,18 @@ async def update_post_full(
 
 @router.patch("/{post_id}", response_model=PostResponse)
 async def update_post_partial(
-    post_id: int, db: Annotated[AsyncSession, Depends(get_db)], post_data: PostUpdate
+    post_id: int, db: Annotated[AsyncSession, Depends(get_db)],current_user:CurrentUser, post_data: PostUpdate
 ):
     result = await db.execute(Select(models.Post).where(models.Post.id == post_id))
     post = result.scalars().first()
     if not post:
         raise HTTPException(
             status_code=status.HTTP_404_NOT_FOUND, detail="Post not found"
+        )
+    if post.user_id != current_user.id:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="Not authorised to update this post",
         )
 
     update_data = post_data.model_dump(exclude_unset=True)
@@ -124,12 +132,17 @@ async def update_post_partial(
 
 #  deleting a post
 @router.delete("/{post_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def delete_post(post_id: int, db: Annotated[AsyncSession, Depends(get_db)]):
+async def delete_post(post_id: int,current_user:CurrentUser, db: Annotated[AsyncSession, Depends(get_db)]):
     result = await db.execute(Select(models.Post).where(models.Post.id == post_id))
     post = result.scalars().first()
     if not post:
         raise HTTPException(
             status_code=status.HTTP_404_NOT_FOUND, detail="Post not found"
+        )
+    if post.user_id != current_user.id:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="Not authorised to delete this post",
         )
 
     await db.delete(post)
